Prototype draw manager.
Consists of:
 * a new inheritable class "gui2::top_level_drawable"
 * a handler for drawing, both in-order and as-required
 * an interface for marking regions to be redrawn.

top_level_drawable should be inherited by any top-level drawable object.
It requires implementation of an "expose" method, which is used to
perform a draw.

The gui2::draw_manager namespace handles drawing, and marking for draw.
"draw_manager::draw()" should be called every vsync frame, and will only
draw whatever is necessary. It returns true if anything was drawn.
Thus the screen only needs flipping/presenting if this returns true.
It will call "expose" on any relevant top_level_drawable. Currently
this is all top-level drawables, but eventually locations should be
tracked or queried.

Widgets are expected to inform the draw manager when they need to be
redrawn. This is done at low level via "draw_manager::invalidate_region",
but more often at high level via the wrapper "widget::queue_draw()".
Assuming the widget has been laid out, it needs only call queue_draw()
whenever its visible state changes.

Animations will require more handling. For now there's a simple handler
that allows a top-level-drawable to register for animation. In this case,
"expose" is called every draw frame. This will likely change, and it just
here to make things work for now.

This is a rough draft, and much of is is likely to shift. But the basic
structure should remain approximately the same.